Updated the way we read 2D data if sinogram exists
Option 'slice' offers the flexibility of reconstructing a 2D slice other than the very centre slice of the beads data. In case the user has not given a specific slice, the default 2D slice is chosen to be the centre. The commit here reads the sinogram file (of which there is only one, and it is of the centre slice) for a faster data loading and for those that do not wish to download the full 3D data.